refactor(text): create types folder
